Fixed bug in dire.control.Reactor
The reactor didn't check right after startup whether it should
already shutdown again. This could lead to a situation where
constant signals that met the abort condition already with
the initial value but never fired an event to the Reactor, would
run forever.

This commit changed the behavior so that the Reactor checks
after startup and after each processed event whether it is time
to shutdown.