Prevent deadlocks in tests using thread pools
Tests using thread pools (that is almost all tests in dire)
have a tendency to deadlock if the number of threads created or
used exceeded the number of cores of the system the tests are
running on.

This fix in Build.scala might solve this issue although I will have
to test it on several machines.