Added and tested function SF.io
This is another helper function to interact with the impure outside
world. Sometimes creating a signal function itself is an IO action.
This happens for instance when creating parts of a user interface. As
soon as all layout stuff is done, we are only interested in the
behavior, which is now wrapped in an IO-action. In order to still
combine it with other SFs, this new helper function can be used.