Reimplemented and documented dire.test.simulate
The processing of every event needs now to be confirmed before
the next event is being fired. This gives much more control over
when a new event is fired. For instance, in a swing-based user interface
just wait until the UI itself has been updated. The rest of the
reactive graph will have followed.