ComboBox.in no longer cached by default
Caching of signal functions can be circumvented in most (probably
even all) cases, so it might be of much less importance than initially
assumed. On the other hand, TypeTags are not thread-safe and might not
be available all the time, therefore the requirement for an
implicit TypeTag when creating ComboBoxes and their behaviors was
removed. Type class IOWidget now provides functions cachedIn and
cachedSf for those occasions where caching seems to be unavoidable.